Change likeCount / dislikeCount field datatype from int to long

+++ src/main/java/io/github/adampyramide/BlogAPI/blogpost/BlogPostResponseDTO.java
@@ -11,7 +11,7 @@ public record BlogPostResponseDTO(
         String title,
         String body,
         LocalDateTime createTime,
-        int likeCount,
-        int dislikeCount
+        long likeCount,
+        long dislikeCount
 
 ) {}
